feat(CozyTheme): Move it into providers folder
BREAKING CHANGE: `CozyTheme` has been moved. Use codemods to deal with the change.
