+++ react/providers/Readme.md
@@ -1,0 +1,1 @@
+<!-- For styleguidist, a .md file is needed in directories -->
